Save empty list when we get no duo_roles parameter in the POST body (#15)
Summary:
When the wordpress settings page for multisite admin is submitted, checkbox-based
parameters with all empty checkboxes do not appear in the POST body. This is correctly handled in
the case of `duo_xmlrpc` where a missing value defaults to 'on' (duo_wordpress.php:474) but was
unhandled in the case of `duo_roles` leading to the behavior that submitting no roles left the roles
value in the database unmodified rather than saving "no roles." This has been correct in this diff.

It would appear this "bug" has existed for quite some time as it still appeared in wordpress 4.3
and I suspect that the way checkboxes are submitted in POST requests it not something that wordpress
changes frequently.

Why would anyone want to submit blank roles?
Ironically this is actually part of our test plan. The reason is that users with "No role for this
site" (aka the role "None") will always get 2FA even with all roles unchecked. If an
administrator wanted 2FA only for users without a role then they would uncheck all boxes.

Co-authored-by: Mark Bishop <mbishop@duosecurity.com>